requirements.txt
paho-mqtt==0.4.94
pjsua==2.2.1
